Several changes required for model2roms to run with very large grids (e.g. the KINO grid of 1600x1600x40) where the arrays can exceed 2 GB. For this reason, double precision has been replaced with REAL(4) in the Fortran scripts. Also added a run script for queue systems such as super computers.

+++ cleanArray.f90
@@ -64,12 +64,12 @@ Module cleanArray
             integer maxpoints, II, JJ, KK, VV, ic, jc, jcc, icc, fill, kc, ff, total
             integer pointsHpos, pointsHneg, pointsVpos, pointsVneg, points, INCREASED, orgVV, orgKK, TRIES
 
-            double precision, dimension(JJ,II) :: dataout,datain
-            double precision, dimension(JJ,II) :: mask
-            double precision, dimension(4,3,maxpoints)  :: foundWeight, foundData
+            REAL(4), dimension(JJ,II) :: dataout,datain
+            REAL(4), dimension(JJ,II) :: mask
+            REAL(4), dimension(4,3,maxpoints)  :: foundWeight, foundData
 
-            double precision counter, hori, vert, sumweights, distance, sodadepth, weight
-            double precision, dimension(JJ,II) :: bathymetry
+            REAL(4) counter, hori, vert, sumweights, distance, sodadepth, weight
+            REAL(4), dimension(JJ,II) :: bathymetry
 
 !f2py intent(in,out,overwrite) dataout
 !f2py intent(in,overwrite) datain, mask, bathymetry
